move BuilderTest into builder_test file
