+++ builder/builder_test.go
@@ -9,6 +9,49 @@ import (
 	"github.com/ulule/loukoum/stmt"
 )
 
+type BuilderTest struct {
+	Name       string
+	Builder    builder.Builder
+	String     string
+	Query      string
+	Args       []interface{}
+	NamedQuery string
+	NamedArgs  map[string]interface{}
+	Failure    func() builder.Builder
+}
+
+func (b *BuilderTest) Run(t *testing.T) {
+	if b.Failure != nil {
+		t.Run("Failure", func(t *testing.T) {
+			require.Panics(t, func() {
+				_ = b.Failure()
+			})
+		})
+	}
+	if b.Builder == nil {
+		return
+	}
+	t.Run("String", func(t *testing.T) {
+		if b.String != "" {
+			require.Equal(t, b.String, b.Builder.String())
+		}
+	})
+	t.Run("Query", func(t *testing.T) {
+		query, args := b.Builder.Query()
+		if b.Query != "" {
+			require.Equal(t, b.Query, query)
+			require.Equal(t, b.Args, args)
+		}
+	})
+	t.Run("NamedQuery", func(t *testing.T) {
+		query, args := b.Builder.NamedQuery()
+		if b.NamedQuery != "" {
+			require.Equal(t, b.NamedQuery, query)
+			require.Equal(t, b.NamedArgs, args)
+		}
+	})
+}
+
 func Failure(is *require.Assertions, callback func() builder.Builder) {
 	is.Panics(func() {
 		_ = callback().String()
